+++ docs/internal/processes/development-guide-relevant-files.md
@@ -172,6 +172,7 @@ This inventory records the checked-in files and directories that the maintainer 
 | `pkg/factory/runtime/dispatch_result_hook_test.go` | Dispatch-hook regression surface | Planner-present synchronous execution, planner-absent async worker-pool submission, missing-runner failures, and replay-tick validation should stay covered here when the behavior being changed is local to `workerPoolDispatchResultHook`, instead of forcing every dispatch-hook regression through the broader factory integration suite. |
 | `pkg/factory/subsystems/subsystem_dispatcher.go` | Dispatcher snapshot-enrichment owner | Dispatcher execution must reattach `d.state` onto raw runtime snapshots before handing them to scheduler/enablement flows, because runtime-only snapshots intentionally omit topology while runtime guards still need normalized transition-to-worker mappings. |
 | `pkg/workers/executor/helpers.go` | Executor-local event and environment diagnostic helper owner | Worker executor event payload shaping, safe command-environment diagnostics, and zero-vs-nonzero exit-code emission now belong with `pkg/workers/executor` so script and workstation orchestration can share one local helper seam without pulling those mechanics back into the stable `pkg/workers` root package. |
+| `pkg/workers/executor/agent/`, `pkg/workers/executor/pool/`, `pkg/workers/executor/script/`, and `pkg/workers/executor/workstation/` | Executor package-size test split seam | Keep exported-surface agent stop-token, agent output-schema, worker-pool, noop-executor, basic script-executor, and workstation timeout coverage in these focused external test packages when executor root tests approach the package file-count cap. Tests that assert package-private script helpers, model-operation binding internals, or workstation context-resolution helpers should remain in root `pkg/workers/executor` and be consolidated by concern instead of exporting production internals for package-size cleanup. |
 | `pkg/workers/process/` and `pkg/workers/command.go` | Worker subprocess boundary split | Subprocess request shaping, environment merging, timeout and exit handling, and platform-specific command execution now belong under `pkg/workers/process`, while `pkg/workers/command.go` stays as the stable compatibility seam for broadly consumed worker command types. Future worker package moves should preserve this pattern: move implementation into the narrow subpackage first, then keep or trim the root boundary deliberately instead of letting execution mechanics drift back into `pkg/workers`. |
 | `pkg/workers/mockworker/` and `pkg/workers/provider_compat.go` | Mock-worker command-runner boundary split | Mock-worker command-runner behavior belongs under `pkg/workers/mockworker`; the root `pkg/workers` package should expose only the compatibility alias needed by service callers. Keep future mock-worker matching, script delegation, and selector tests adjacent to the mockworker package instead of moving implementation or same-package tests back into the root workers package. |
 | `pkg/workers/prompting/template_fields.go` | Workstation parameterized field-resolution owner | Working-directory, environment-variable, and worktree template resolution for workstation execution should stay centralized here so runtime cleanup work updates one explicit helper contract instead of reintroducing call-shape shims or duplicating template evaluation in `pkg/workers/executor/workstation_executor.go` or tests. |
