+++ pkg/workers/executor/agent/agent_stop_token_test.go
@@ -1,19 +1,60 @@
-package executor
+package agent_test
 
 import (
 	"context"
 	"testing"
 
 	"github.com/portpowered/infinite-you/pkg/interfaces"
+	"github.com/portpowered/infinite-you/pkg/testutil/runtimefixtures"
+	executorpkg "github.com/portpowered/infinite-you/pkg/workers/executor"
 )
 
+type agentMockProvider struct {
+	response  interfaces.InferenceResponse
+	callCount int
+	lastReq   interfaces.ProviderInferenceRequest
+}
+
+func (m *agentMockProvider) Infer(_ context.Context, req interfaces.ProviderInferenceRequest) (interfaces.InferenceResponse, error) {
+	m.lastReq = req
+	m.callCount++
+	return m.response, nil
+}
+
+func testAgentRequest(dispatch interfaces.WorkDispatch, opts ...func(*interfaces.WorkstationExecutionRequest)) interfaces.WorkstationExecutionRequest {
+	req := interfaces.WorkstationExecutionRequest{
+		Dispatch:        interfaces.CloneWorkDispatch(dispatch),
+		WorkerType:      dispatch.WorkerType,
+		WorkstationType: dispatch.WorkstationName,
+		ProjectID:       dispatch.ProjectID,
+		InputTokens:     append([]any(nil), dispatch.InputTokens...),
+	}
+	for _, opt := range opts {
+		opt(&req)
+	}
+	return req
+}
+
+func withAgentPrompts(systemPrompt, userMessage string) func(*interfaces.WorkstationExecutionRequest) {
+	return func(req *interfaces.WorkstationExecutionRequest) {
+		req.SystemPrompt = systemPrompt
+		req.UserMessage = userMessage
+	}
+}
+
+func withAgentOutputSchema(schema string) func(*interfaces.WorkstationExecutionRequest) {
+	return func(req *interfaces.WorkstationExecutionRequest) {
+		req.OutputSchema = schema
+	}
+}
+
 func TestAgentExecutor_StopTokenControlsOutcome(t *testing.T) {
-	runtimeCfg := staticRuntimeConfig{
+	runtimeCfg := runtimefixtures.RuntimeConfigLookupFixture{
 		Workers: map[string]*interfaces.WorkerConfig{
 			"worker-a": {Model: "test-model", StopToken: "COMPLETE"},
 		},
 	}
-	executor := NewAgentExecutor(
+	executor := executorpkg.NewAgentExecutor(
 		runtimeCfg,
 		&agentMockProvider{response: interfaces.InferenceResponse{Content: "Work done. COMPLETE"}},
 	)
@@ -33,7 +74,7 @@ func TestAgentExecutor_StopTokenControlsOutcome(t *testing.T) {
 		t.Fatalf("Outcome = %s, want %s", result.Outcome, interfaces.OutcomeAccepted)
 	}
 
-	executor = NewAgentExecutor(
+	executor = executorpkg.NewAgentExecutor(
 		runtimeCfg,
 		&agentMockProvider{response: interfaces.InferenceResponse{Content: "Still working"}},
 	)
@@ -52,7 +93,7 @@ func TestAgentExecutor_StopTokenControlsOutcome(t *testing.T) {
 		t.Fatalf("Outcome = %s, want %s", result.Outcome, interfaces.OutcomeRejected)
 	}
 
-	executor = NewAgentExecutor(
+	executor = executorpkg.NewAgentExecutor(
 		runtimeCfg,
 		&agentMockProvider{response: interfaces.InferenceResponse{Content: "Still iterating\n<CONTINUE>"}},
 	)
@@ -74,7 +115,7 @@ func TestAgentExecutor_StopTokenControlsOutcome(t *testing.T) {
 
 func TestAgentExecutor_StopTokenComesFromRuntimeConfigWithoutDispatchState(t *testing.T) {
 	provider := &agentMockProvider{response: interfaces.InferenceResponse{Content: "Work done. COMPLETE"}}
-	executor := NewAgentExecutor(staticRuntimeConfig{
+	executor := executorpkg.NewAgentExecutor(runtimefixtures.RuntimeConfigLookupFixture{
 		Workers: map[string]*interfaces.WorkerConfig{
 			"worker-a": {Model: "test-model", StopToken: "COMPLETE"},
 		},
@@ -99,12 +140,12 @@ func TestAgentExecutor_StopTokenComesFromRuntimeConfigWithoutDispatchState(t *te
 func TestAgentExecutor_RuntimeStopTokenChangesAffectSubsequentDispatches(t *testing.T) {
 	provider := &agentMockProvider{response: interfaces.InferenceResponse{Content: "Work done. COMPLETE"}}
 	workerDef := &interfaces.WorkerConfig{Model: "test-model", StopToken: "COMPLETE"}
-	runtimeCfg := staticRuntimeConfig{
+	runtimeCfg := runtimefixtures.RuntimeConfigLookupFixture{
 		Workers: map[string]*interfaces.WorkerConfig{
 			"worker-a": workerDef,
 		},
 	}
-	executor := NewAgentExecutor(runtimeCfg, provider)
+	executor := executorpkg.NewAgentExecutor(runtimeCfg, provider)
 
 	dispatch := interfaces.WorkDispatch{
 		DispatchID:   "d-1",
@@ -141,7 +182,7 @@ func TestAgentExecutor_RuntimeStopTokenChangesAffectSubsequentDispatches(t *test
 
 func TestAgentExecutor_ResolvesWorkerConfigPerDispatch(t *testing.T) {
 	provider := &agentMockProvider{response: interfaces.InferenceResponse{Content: "done"}}
-	executor := NewAgentExecutor(staticRuntimeConfig{
+	executor := executorpkg.NewAgentExecutor(runtimefixtures.RuntimeConfigLookupFixture{
 		Workers: map[string]*interfaces.WorkerConfig{
 			"worker-a": {Model: "model-a", ModelProvider: "claude"},
 			"worker-b": {Model: "model-b", ModelProvider: "codex"},
@@ -182,5 +223,66 @@ func TestAgentExecutor_ResolvesWorkerConfigPerDispatch(t *testing.T) {
 	}
 	if provider.lastReq.Model != "model-b" || provider.lastReq.ModelProvider != "codex" {
 		t.Fatalf("second request = %#v", provider.lastReq)
+	}
+}
+
+func TestAgentExecutor_OutputSchemaSuccess_KeepsRawOutput(t *testing.T) {
+	provider := &agentMockProvider{response: interfaces.InferenceResponse{Content: `{"work_id":"w-1","tags":{"result":"done"}}`}}
+	executor := executorpkg.NewAgentExecutor(runtimefixtures.RuntimeConfigLookupFixture{
+		Workers: map[string]*interfaces.WorkerConfig{
+			"worker-a": {Model: "test-model"},
+		},
+	}, provider)
+
+	result, err := executor.Execute(context.Background(), testAgentRequest(
+		interfaces.WorkDispatch{
+			DispatchID:   "d-1",
+			TransitionID: "t-1",
+			WorkerType:   "worker-a",
+		},
+		withAgentPrompts("sys", "msg"),
+		withAgentOutputSchema(`{"type":"object"}`),
+	))
+	if err != nil {
+		t.Fatalf("unexpected error: %v", err)
+	}
+
+	if result.Outcome != interfaces.OutcomeAccepted {
+		t.Fatalf("Outcome = %s, want %s", result.Outcome, interfaces.OutcomeAccepted)
+	}
+	if result.Output != `{"work_id":"w-1","tags":{"result":"done"}}` {
+		t.Fatalf("Output = %q", result.Output)
+	}
+}
+
+func TestAgentExecutor_OutputSchemaParseFailure_ReturnsFailedResult(t *testing.T) {
+	provider := &agentMockProvider{response: interfaces.InferenceResponse{Content: "not valid json at all"}}
+	executor := executorpkg.NewAgentExecutor(runtimefixtures.RuntimeConfigLookupFixture{
+		Workers: map[string]*interfaces.WorkerConfig{
+			"worker-a": {Model: "test-model"},
+		},
+	}, provider)
+
+	result, err := executor.Execute(context.Background(), testAgentRequest(
+		interfaces.WorkDispatch{
+			DispatchID:   "d-1",
+			TransitionID: "t-1",
+			WorkerType:   "worker-a",
+		},
+		withAgentPrompts("sys", "msg"),
+		withAgentOutputSchema(`{"type":"object"}`),
+	))
+	if err != nil {
+		t.Fatalf("unexpected error: %v", err)
+	}
+
+	if result.Outcome != interfaces.OutcomeFailed {
+		t.Fatalf("Outcome = %s, want %s", result.Outcome, interfaces.OutcomeFailed)
+	}
+	if result.Error == "" {
+		t.Fatal("expected parse error")
+	}
+	if result.Output != "not valid json at all" {
+		t.Fatalf("Output = %q, want raw response", result.Output)
 	}
 }
